moved custom hook useBrandOptions to its own file
