+++ src/components/ProductsListFilters.js
@@ -1,36 +1,12 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { Form, Dropdown } from "semantic-ui-react";
 import { useSelector, useDispatch } from "react-redux";
-import {
-  selectStatus,
-  selectActiveBrand,
-  selectBrandOptions,
-} from "store/products/selectors";
-import {
-  fetchBrands,
-  fetchBrandProducts,
-  filterByName,
-} from "store/products/slice";
-import { STATUS } from "store/products/constants";
-
-function useBrandOptions() {
-  const dispatch = useDispatch();
-  const status = useSelector(selectStatus);
-  const brandOptions = useSelector(selectBrandOptions);
-  const isEmpty = status === STATUS.INITIAL && brandOptions.length === 0;
-  useEffect(
-    function loadBrandNames() {
-      if (isEmpty) {
-        dispatch(fetchBrands());
-      }
-    },
-    [isEmpty, dispatch]
-  );
-  return brandOptions;
-}
+import { selectActiveBrand } from "store/products/selectors";
+import { fetchBrandProducts, filterByName } from "store/products/slice";
+import useBrandOptions from "hooks/useBrandOptions";
 
 export default function ProductsListFilters() {
-  const brandOptions = useBrandOptions(selectBrandOptions);
+  const brandOptions = useBrandOptions();
   const dispatch = useDispatch();
   const brand = useSelector(selectActiveBrand);
   const selectBrand = (_, data) => {
